fix(perry-ext-net): keep lib.rs under the 2000-line file-size gate (#2131)
Move `drain_once_listeners` into the `lifecycle` submodule (where the
once-flag side-table also lives) and qualify `HashSet` inside the
`statics` module so `lib.rs` no longer needs `use std::collections::HashSet`.
Drops lib.rs from 2003 → 1974 lines, clearing the
`scripts/check_file_size.sh` CI gate.
